agrego condición  de produccióntambién a save dpara eliminar files; modelos del IMSS

+++ respond/models.py
@@ -12,7 +12,7 @@ from respond.data_file_mixins.utils_mix import DataUtilsMix
 from inai.models import Petition, set_upload_path, PetitionFileControl, MonthRecord, WeekRecord
 from respond.reply_file_mixins.process_mix import ReplyFileMix
 
-CAN_DELETE_AWS_STORAGE_FILES = getattr(
+can_delete_s3 = getattr(
     settings, "CAN_DELETE_AWS_STORAGE_FILES", False)
 
 
@@ -41,7 +41,7 @@ class ReplyFile(models.Model, ReplyFileMix):
         default=False, verbose_name="Contiene los datos")
 
     def save(self, *args, **kwargs):
-        if self.pk:
+        if self.pk and can_delete_s3:
             try:
                 old_instance = ReplyFile.objects.get(pk=self.pk)
                 if old_instance.file and old_instance.file != self.file:
@@ -58,7 +58,7 @@ class ReplyFile(models.Model, ReplyFileMix):
             raise Exception(
                 "No se puede eliminar un archivo con datos insertados")
 
-        if CAN_DELETE_AWS_STORAGE_FILES:
+        if can_delete_s3:
             self.file.delete(save=False)
         else:
             print("No se pueden borrar archivos en AWS")
@@ -127,7 +127,7 @@ class DataFile(models.Model, ExploreMix, DataUtilsMix, ExtractorsMix):
     notes = models.TextField(blank=True, null=True)
 
     def save(self, *args, **kwargs):
-        if self.pk:
+        if self.pk and can_delete_s3:
             try:
                 old_instance = DataFile.objects.get(pk=self.pk)
                 if old_instance.file and old_instance.file != self.file:
@@ -144,7 +144,7 @@ class DataFile(models.Model, ExploreMix, DataUtilsMix, ExtractorsMix):
             raise Exception(
                 "No se puede eliminar un archivo con datos insertados")
 
-        if CAN_DELETE_AWS_STORAGE_FILES:
+        if can_delete_s3:
             self.file.delete(save=False)
         else:
             print("No se pueden borrar archivos en AWS")
@@ -295,7 +295,7 @@ class SheetFile(models.Model):
     # all_results = JSONField(blank=True, null=True)
 
     def save(self, *args, **kwargs):
-        if self.pk:
+        if self.pk and can_delete_s3:
             try:
                 old_instance = SheetFile.objects.get(pk=self.pk)
                 if old_instance.file and old_instance.file != self.file:
@@ -310,7 +310,7 @@ class SheetFile(models.Model):
         if some_inserted:
             raise Exception(
                 "No se puede eliminar un archivo con datos insertados")
-        if CAN_DELETE_AWS_STORAGE_FILES:
+        if can_delete_s3:
             self.file.delete(save=False)
         else:
             print("No se pueden borrar archivos en AWS")
@@ -468,9 +468,10 @@ class TableFile(models.Model):
         return "%s %s" % (self.collection, self.lap_sheet)
 
     def save(self, *args, **kwargs):
-        if self.pk:
+        if self.pk and can_delete_s3:
             try:
                 old_instance = TableFile.objects.get(pk=self.pk)
+
                 if old_instance.file and old_instance.file != self.file:
                     old_instance.file.delete(save=False)
             except TableFile.DoesNotExist:
@@ -479,7 +480,7 @@ class TableFile(models.Model):
         super().save(*args, **kwargs)
 
     def delete(self, *args, **kwargs):
-        if CAN_DELETE_AWS_STORAGE_FILES:
+        if can_delete_s3:
             self.file.delete(save=False)
         else:
             print("No se pueden borrar archivos en AWS")
